Made product page component reusable
